+++ docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/docs
 
+## 2.6.34
+
+### Patch Changes
+
+- Bugfixes
+
+## 2.6.33
+
+## 2.6.32
+
 ## 2.6.31
 
 ### Patch Changes