+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/stack-shared
 
+## 2.6.34
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-sc@2.6.34
+
+## 2.6.33
+
+### Patch Changes
+
+- @stackframe/stack-sc@2.6.33
+
+## 2.6.32
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-sc@2.6.32
+
 ## 2.6.31
 
 ### Patch Changes